+++ spec/open_meteo/forecast_spec.rb
@@ -31,7 +31,7 @@ RSpec.describe OpenMeteo::Forecast do
       end
     end
 
-    context "when the location and the varialbes are valid" do
+    context "when the location and the variables are valid" do
       let(:variables) { { hourly: %i[apparent_temperature], daily: %i[apparent_temperature_max] } }
 
       before do