+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add models to forecast variables
 - Add minutely_15 to forecast variables
 - Add minutely_15 to forecast entity
+- Add DWD Icon forecast model
+- Add integration test with VCR
 
 ## [0.2.0] - 2023-11-08